+++ composeApp/src/commonMain/kotlin/kmp/fbk/kmpartgallery/reusable_ui_compomenents/search/ISearchViewModel.kt
@@ -1,4 +1,4 @@
-package kmp.fbk.kmpartgallery.features.listscreen
+package kmp.fbk.kmpartgallery.reusable_ui_compomenents.search
 
 interface ISearchViewModel {
     fun onSearchTextChange(query: String, additionalArgs: String? = null)
